Reorganization and refactoring
